requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
email-validator
python-multipart
requests
fpdf
Pillow
numpy
asyncpg
python-dotenv
tensorflow
#tflite_runtime===2.13.0
#tflite